use static event page, refactor eventDataPage
