+++ frontend/shared/src/components/hocs/initLocale.tsx
@@ -10,9 +10,7 @@ const initLocale =
     const { locale } = router;
     const { i18n } = useTranslation();
     React.useEffect(() => {
-      if (i18n?.changeLanguage) {
-        void i18n.changeLanguage(locale ?? DEFAULT_LANGUAGE);
-      }
+      void i18n.changeLanguage(locale ?? DEFAULT_LANGUAGE);
     }, [i18n, locale]);
     return <WrappedComponent {...props} />;
   };
