+++ frontend/tet/youth/src/eventDataPage/PlaceView.tsx
